Handle case where Dyn returns an empty CSV
Dyn had a bug in its QPS-reporting code which was resulting in returning
a CSV file with no data (only header). Trying to extractSecondLastQps
on that data led to a panic trying to access `raw[keys[len(keys)-2]]`
on a nil map.

I was torn as to whether to put the error checking in getQpsReport()
or in extractSecondLastQps and am happy to accept feedback :-)