Ask for 20 minutes of QPS data, not 15
There is a Dyn API bug(?) in which asking for 15 minutes of
every-5-minute data will only return the current period of data. We
have a support ticket open with them, but asking for 20 minutes of data
appears to reliably return at least one complete period plus the current
period, so let's do that.